psi: Move PF_MEMSTALL out of task->flags
The task->flags is a 32-bits flag, in which 31 bits have already been
consumed. So it is hardly to introduce other new per process flag.
Currently there're still enough spaces in the bit-field section of
task_struct, so we can define the memstall state as a single bit in
task_struct instead.
This patch also removes an out-of-date comment pointed by Matthew.

Suggested-by: Johannes Weiner <hannes@cmpxchg.org>
Signed-off-by: Yafang Shao <laoar.shao@gmail.com>
Change-Id: I0066a4d19b7578b664aa5b3bd345eed2494c3ce3
Signed-off-by: Jason Edson <jaysonedson@gmail.com>
Signed-off-by: DennySPb <dennyspb@gmail.com>
Signed-off-by: Little-W <1405481963@qq.com>