kprofiles: use cpu ang gpu load and frequncy to control kprofile auto configs
Use two functions to detect gpu freq and cpu freq. Remove the limitation of cpu freq when high freq is continuously requested for a period.
If gpu freq stay high for a period of time, we assume that a game is
running, and let kp_active_mode = 4. If gpu freq is high just for a short
period, we assume that a performance consuming task is running (like the
animator of miui), then kp_active_mode = 3 to enable performance mode
temporarily.

* kp_active_mode() == 4 : gaming
* kp_active_mode() == 3 : performance
* kp_active_mode() == 2 : balance
* kp_active_mode() == 1 : powersave